Update windows paths in Readme
Documents and settings folder doesn't exist since Windows Vista (2007). I guess this macro isn't targeted for XP or NT users only 😂

Updated docs to the current correct paths. If you don't like the `%appdata%` version, you can delete it, but I always access appdata folder this way, it's quicker and works regardless of username, just like `~` in *nix, but it can be confusing for users not familiar with environment variables.